+++ test/types.js
@@ -814,3 +814,16 @@ describe('int16Array', function () {
         assert.equal(result, false);
     });
 });
+
+describe('uint16Array', function () {
+    it('should be return true', function () {
+        var result = be.uint16Array(new Uint16Array({}));
+        console.log(result);
+        assert.equal(result, true);
+    });
+    it('should be return false', function () {
+        var result = be.uint16Array({});
+        console.log(result);
+        assert.equal(result, false);
+    });
+});