+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - es6 `be.uint8Array`, Uint8Array object assertion
 - es6 `be.uint8ClampedArray`, Uint8ClampedArray object assertion
 - es6 `be.int16Array`, Int16Array object assertion
+- es6 `be.uint16Array`, Uint16Array object assertion
 
 ## [1.4.1] - 2017-07-10